test/yusun/my_mlops/airflow/requirements.txt
requests
boto3
pendulum==2.1.2
python-magic
